move OpeningProof and open to evaluation_proof.rs
