+++ poly-commitment/src/lib.rs
@@ -2,5 +2,6 @@ pub mod chunked;
 mod combine;
 pub mod commitment;
 pub mod srs;
+pub mod evaluation_proof;
 
 pub use commitment::PolyComm;
